Refactor `icrate` mutable counterpart files
