+++ crates/icrate/src/Foundation/additions/attributed_string.rs
@@ -15,32 +15,45 @@ unsafe impl Send for NSAttributedString {}
 impl UnwindSafe for NSAttributedString {}
 impl RefUnwindSafe for NSAttributedString {}
 
-extern_methods!(
-    /// Creating attributed strings.
-    unsafe impl NSAttributedString {
-        /// Creates a new attributed string from the given string and attributes.
-        ///
-        /// The attributes are associated with every UTF-16 code unit in the
-        /// string.
-        ///
-        /// # Safety
-        ///
-        /// The attributes must be valid.
-        #[doc(alias = "initWithString:")]
-        #[cfg(feature = "Foundation_NSDictionary")]
-        #[cfg(feature = "Foundation_NSString")]
-        pub unsafe fn new_with_attributes(
-            string: &Foundation::NSString,
-            attributes: &Foundation::NSDictionary<NSAttributedStringKey, Object>,
-        ) -> Id<Self> {
-            unsafe { Self::initWithString_attributes(Self::alloc(), string, Some(attributes)) }
-        }
-
-        /// Creates a new attributed string without any attributes.
-        #[doc(alias = "initWithString:")]
-        #[cfg(feature = "Foundation_NSString")]
-        pub fn from_nsstring(string: &Foundation::NSString) -> Id<Self> {
-            Self::initWithString(Self::alloc(), string)
-        }
+impl NSAttributedString {
+    /// Creates a new attributed string from the given string and attributes.
+    ///
+    /// The attributes are associated with every UTF-16 code unit in the
+    /// string.
+    ///
+    /// # Safety
+    ///
+    /// The attributes must be valid.
+    #[doc(alias = "initWithString:")]
+    #[cfg(feature = "Foundation_NSDictionary")]
+    #[cfg(feature = "Foundation_NSString")]
+    pub unsafe fn new_with_attributes(
+        string: &Foundation::NSString,
+        attributes: &Foundation::NSDictionary<NSAttributedStringKey, Object>,
+    ) -> Id<Self> {
+        unsafe { Self::initWithString_attributes(Self::alloc(), string, Some(attributes)) }
     }
-);
+
+    /// Creates a new attributed string without any attributes.
+    #[doc(alias = "initWithString:")]
+    #[cfg(feature = "Foundation_NSString")]
+    pub fn from_nsstring(string: &Foundation::NSString) -> Id<Self> {
+        Self::initWithString(Self::alloc(), string)
+    }
+}
+
+#[cfg(feature = "Foundation_NSMutableAttributedString")]
+impl Foundation::NSMutableAttributedString {
+    // TODO: new_with_attributes
+
+    #[doc(alias = "initWithString:")]
+    #[cfg(feature = "Foundation_NSString")]
+    pub fn from_nsstring(string: &Foundation::NSString) -> Id<Self> {
+        Self::initWithString(Self::alloc(), string)
+    }
+
+    #[doc(alias = "initWithAttributedString:")]
+    pub fn from_attributed_nsstring(attributed_string: &NSAttributedString) -> Id<Self> {
+        Self::initWithAttributedString(Self::alloc(), attributed_string)
+    }
+}
