+++ crates/icrate/src/Foundation/additions/mod.rs
@@ -21,12 +21,6 @@ mod enumerator;
 mod error;
 mod exception;
 mod geometry;
-mod mutable_array;
-mod mutable_attributed_string;
-mod mutable_data;
-mod mutable_dictionary;
-mod mutable_set;
-mod mutable_string;
 mod number;
 mod process_info;
 mod range;
