--- crates/icrate/src/Foundation/additions/mutable_attributed_string.rs
@@ -1,19 +1,0 @@
-#![cfg(feature = "Foundation_NSMutableAttributedString")]
-use crate::common::*;
-use crate::Foundation::{self, NSAttributedString, NSMutableAttributedString};
-
-/// Creating mutable attributed strings.
-impl NSMutableAttributedString {
-    // TODO: new_with_attributes
-
-    #[doc(alias = "initWithString:")]
-    #[cfg(feature = "Foundation_NSString")]
-    pub fn from_nsstring(string: &Foundation::NSString) -> Id<Self> {
-        Self::initWithString(Self::alloc(), string)
-    }
-
-    #[doc(alias = "initWithAttributedString:")]
-    pub fn from_attributed_nsstring(attributed_string: &NSAttributedString) -> Id<Self> {
-        Self::initWithAttributedString(Self::alloc(), attributed_string)
-    }
-}
