+++ crates/icrate/src/Foundation/additions/set.rs
@@ -8,57 +8,195 @@ use objc2::mutability::IsRetainable;
 
 use super::util;
 use crate::common::*;
+#[cfg(feature = "Foundation_NSMutableSet")]
+use crate::Foundation::NSMutableSet;
 use crate::Foundation::{self, NSSet};
+
+impl<T: Message> NSSet<T> {
+    /// Creates an [`NSSet`] from a vector.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use icrate::Foundation::{NSSet, NSString};
+    ///
+    /// let strs = ["one", "two", "three"].map(NSString::from_str).to_vec();
+    /// let set = NSSet::from_vec(strs);
+    /// ```
+    pub fn from_vec(mut vec: Vec<Id<T>>) -> Id<Self> {
+        let len = vec.len();
+        let ptr = util::id_ptr_cast(vec.as_mut_ptr());
+        // SAFETY: Same as `NSArray::from_vec`.
+        unsafe { Self::initWithObjects_count(Self::alloc(), ptr, len) }
+    }
+
+    /// Creates an [`NSSet`] from a slice of `Id`s.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use icrate::Foundation::{NSSet, NSString};
+    ///
+    /// let strs = ["one", "two", "three"].map(NSString::from_str);
+    /// let set = NSSet::from_id_slice(&strs);
+    /// ```
+    pub fn from_id_slice(slice: &[Id<T>]) -> Id<Self>
+    where
+        T: IsIdCloneable,
+    {
+        let len = slice.len();
+        let ptr = util::id_ptr_cast_const(slice.as_ptr());
+        // SAFETY: Same as `NSArray::from_id_slice`
+        unsafe { Self::initWithObjects_count(Self::alloc(), ptr, len) }
+    }
+
+    pub fn from_slice(slice: &[&T]) -> Id<Self>
+    where
+        T: IsRetainable,
+    {
+        let len = slice.len();
+        let ptr = util::ref_ptr_cast_const(slice.as_ptr());
+        // SAFETY: Same as `NSArray::from_slice`.
+        unsafe { Self::initWithObjects_count(Self::alloc(), ptr, len) }
+    }
+
+    /// Returns a [`Vec`] containing the set's elements.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use icrate::Foundation::{NSMutableString, NSSet};
+    ///
+    /// let strs = vec![
+    ///     NSMutableString::from_str("one"),
+    ///     NSMutableString::from_str("two"),
+    ///     NSMutableString::from_str("three"),
+    /// ];
+    /// let set = NSSet::from_vec(strs);
+    /// let vec = set.to_vec();
+    /// assert_eq!(vec.len(), 3);
+    /// ```
+    pub fn to_vec(&self) -> Vec<&T> {
+        self.into_iter().collect()
+    }
+
+    #[doc(alias = "getObjects:range:")]
+    pub fn to_vec_retained(&self) -> Vec<Id<T>>
+    where
+        T: IsIdCloneable,
+    {
+        // SAFETY: The objects are stored in the set
+        self.into_iter()
+            .map(|obj| unsafe { util::collection_retain_id(obj) })
+            .collect()
+    }
+
+    /// Returns an [`NSArray`] containing the set's elements, or an empty
+    /// array if the set is empty.
+    ///
+    /// [`NSArray`]: crate::Foundation::NSArray
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use icrate::Foundation::{NSNumber, NSSet, NSString};
+    ///
+    /// let nums = [1, 2, 3];
+    /// let set = NSSet::from_id_slice(&nums.map(NSNumber::new_i32));
+    ///
+    /// assert_eq!(set.to_array().len(), 3);
+    /// assert!(set.to_array().iter().all(|i| nums.contains(&i.as_i32())));
+    /// ```
+    #[doc(alias = "allObjects")]
+    #[cfg(feature = "Foundation_NSArray")]
+    pub fn to_array(&self) -> Id<Foundation::NSArray<T>>
+    where
+        T: IsIdCloneable,
+    {
+        // SAFETY: The `T: IsIdCloneable` bound ensures that it is safe to
+        // create what is effectively a copy from an `&self` reference.
+        //
+        // Could be implemented as:
+        //    NSArray::from_vec(self.to_vec_retained())
+        unsafe { self.allObjects() }
+    }
+}
+
+#[cfg(feature = "Foundation_NSMutableSet")]
+impl<T: Message> NSMutableSet<T> {
+    /// Creates an [`NSMutableSet`] from a vector.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use icrate::Foundation::{NSMutableSet, NSString};
+    ///
+    /// let strs = ["one", "two", "three"].map(NSString::from_str).to_vec();
+    /// let set = NSMutableSet::from_vec(strs);
+    /// ```
+    pub fn from_vec(mut vec: Vec<Id<T>>) -> Id<Self> {
+        let len = vec.len();
+        let ptr = util::id_ptr_cast(vec.as_mut_ptr());
+        // SAFETY: Same as `NSArray::from_vec`.
+        unsafe { Self::initWithObjects_count(Self::alloc(), ptr, len) }
+    }
+
+    /// Creates an [`NSMutableSet`] from a slice of `Id`s.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use icrate::Foundation::{NSMutableSet, NSString};
+    ///
+    /// let strs = ["one", "two", "three"].map(NSString::from_str);
+    /// let set = NSMutableSet::from_id_slice(&strs);
+    /// ```
+    pub fn from_id_slice(slice: &[Id<T>]) -> Id<Self>
+    where
+        T: IsIdCloneable,
+    {
+        let len = slice.len();
+        let ptr = util::id_ptr_cast_const(slice.as_ptr());
+        // SAFETY: Same as `NSArray::from_id_slice`
+        unsafe { Self::initWithObjects_count(Self::alloc(), ptr, len) }
+    }
+
+    pub fn from_slice(slice: &[&T]) -> Id<Self>
+    where
+        T: IsRetainable,
+    {
+        let len = slice.len();
+        let ptr = util::ref_ptr_cast_const(slice.as_ptr());
+        // SAFETY: Same as `NSArray::from_slice`.
+        unsafe { Self::initWithObjects_count(Self::alloc(), ptr, len) }
+    }
+
+    /// Returns a [`Vec`] containing the set's elements, consuming the set.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use icrate::Foundation::{NSMutableSet, NSMutableString};
+    ///
+    /// let strs = vec![
+    ///     NSMutableString::from_str("one"),
+    ///     NSMutableString::from_str("two"),
+    ///     NSMutableString::from_str("three"),
+    /// ];
+    /// let set = NSMutableSet::from_vec(strs);
+    /// let vec = NSMutableSet::into_vec(set);
+    /// assert_eq!(vec.len(), 3);
+    /// ```
+    pub fn into_vec(set: Id<Self>) -> Vec<Id<T>> {
+        // SAFETY: Same as `NSMutableArray::into_vec`
+        set.into_iter()
+            .map(|obj| unsafe { util::mutable_collection_retain_removed_id(obj) })
+            .collect()
+    }
+}
 
 extern_methods!(
     unsafe impl<T: Message> NSSet<T> {
-        /// Creates an [`NSSet`] from a vector.
-        ///
-        /// # Examples
-        ///
-        /// ```
-        /// use icrate::Foundation::{NSSet, NSString};
-        ///
-        /// let strs = ["one", "two", "three"].map(NSString::from_str).to_vec();
-        /// let set = NSSet::from_vec(strs);
-        /// ```
-        pub fn from_vec(mut vec: Vec<Id<T>>) -> Id<Self> {
-            let len = vec.len();
-            let ptr = util::id_ptr_cast(vec.as_mut_ptr());
-            // SAFETY: Same as `NSArray::from_vec`.
-            unsafe { Self::initWithObjects_count(Self::alloc(), ptr, len) }
-        }
-
-        /// Creates an [`NSSet`] from a slice of `Id`s.
-        ///
-        /// # Examples
-        ///
-        /// ```
-        /// use icrate::Foundation::{NSSet, NSString};
-        ///
-        /// let strs = ["one", "two", "three"].map(NSString::from_str);
-        /// let set = NSSet::from_id_slice(&strs);
-        /// ```
-        pub fn from_id_slice(slice: &[Id<T>]) -> Id<Self>
-        where
-            T: IsIdCloneable,
-        {
-            let len = slice.len();
-            let ptr = util::id_ptr_cast_const(slice.as_ptr());
-            // SAFETY: Same as `NSArray::from_id_slice`
-            unsafe { Self::initWithObjects_count(Self::alloc(), ptr, len) }
-        }
-
-        pub fn from_slice(slice: &[&T]) -> Id<Self>
-        where
-            T: IsRetainable,
-        {
-            let len = slice.len();
-            let ptr = util::ref_ptr_cast_const(slice.as_ptr());
-            // SAFETY: Same as `NSArray::from_slice`.
-            unsafe { Self::initWithObjects_count(Self::alloc(), ptr, len) }
-        }
-
         /// Returns the number of elements in the set.
         ///
         /// # Examples
@@ -127,72 +265,11 @@ extern_methods!(
                 Foundation::NSEnumerator2::from_ptr(result)
             }
         }
-
-        /// Returns a [`Vec`] containing the set's elements.
-        ///
-        /// # Examples
-        ///
-        /// ```
-        /// use icrate::Foundation::{NSMutableString, NSSet};
-        ///
-        /// let strs = vec![
-        ///     NSMutableString::from_str("one"),
-        ///     NSMutableString::from_str("two"),
-        ///     NSMutableString::from_str("three"),
-        /// ];
-        /// let set = NSSet::from_vec(strs);
-        /// let vec = set.to_vec();
-        /// assert_eq!(vec.len(), 3);
-        /// ```
-        pub fn to_vec(&self) -> Vec<&T> {
-            self.into_iter().collect()
-        }
-
-        #[doc(alias = "getObjects:range:")]
-        pub fn to_vec_retained(&self) -> Vec<Id<T>>
-        where
-            T: IsIdCloneable,
-        {
-            // SAFETY: The objects are stored in the set
-            self.into_iter()
-                .map(|obj| unsafe { util::collection_retain_id(obj) })
-                .collect()
-        }
-
-        /// Returns an [`NSArray`] containing the set's elements, or an empty
-        /// array if the set is empty.
-        ///
-        /// [`NSArray`]: crate::Foundation::NSArray
-        ///
-        /// # Examples
-        ///
-        /// ```
-        /// use icrate::Foundation::{NSNumber, NSSet, NSString};
-        ///
-        /// let nums = [1, 2, 3];
-        /// let set = NSSet::from_id_slice(&nums.map(NSNumber::new_i32));
-        ///
-        /// assert_eq!(set.to_array().len(), 3);
-        /// assert!(set.to_array().iter().all(|i| nums.contains(&i.as_i32())));
-        /// ```
-        #[doc(alias = "allObjects")]
-        #[cfg(feature = "Foundation_NSArray")]
-        pub fn to_array(&self) -> Id<Foundation::NSArray<T>>
-        where
-            T: IsIdCloneable,
-        {
-            // SAFETY: The `T: IsIdCloneable` bound ensures that it is safe to
-            // create what is effectively a copy from an `&self` reference.
-            //
-            // Could be implemented as:
-            //    NSArray::from_vec(self.to_vec_retained())
-            unsafe { self.allObjects() }
-        }
     }
 
-    // We're explicit about `T` being `PartialEq` for these methods because the
-    // set compares the input value(s) with elements in the set
-    // For comparison: Rust's HashSet requires similar methods to be `Hash` + `Eq`
+    // We're explicit about `T` being `PartialEq` for these methods because
+    // the set compares the input value(s) with elements in the set. For
+    // comparison: Rust's HashSet requires similar methods to be `Hash` + `Eq`
     unsafe impl<T: Message + PartialEq> NSSet<T> {
         /// Returns `true` if the set contains a value.
         ///
@@ -256,8 +333,9 @@ extern_methods!(
         /// assert!(!set2.is_subset(&set1));
         /// ```
         #[doc(alias = "isSubsetOfSet:")]
-        #[method(isSubsetOfSet:)]
-        pub fn is_subset(&self, other: &NSSet<T>) -> bool;
+        pub fn is_subset(&self, other: &NSSet<T>) -> bool {
+            unsafe { self.isSubsetOfSet(other) }
+        }
 
         /// Returns `true` if the set is a superset of another, i.e., `self`
         /// contains at least all the values in `other`.
@@ -277,9 +355,6 @@ extern_methods!(
             other.is_subset(self)
         }
 
-        #[method(intersectsSet:)]
-        fn intersects_set(&self, other: &NSSet<T>) -> bool;
-
         /// Returns `true` if `self` has no elements in common with `other`.
         ///
         /// # Examples
@@ -296,12 +371,64 @@ extern_methods!(
         /// assert!(set2.is_disjoint(&set3));
         /// ```
         pub fn is_disjoint(&self, other: &NSSet<T>) -> bool {
-            !self.intersects_set(other)
+            !unsafe { self.intersectsSet(other) }
         }
     }
 );
 
+#[cfg(feature = "Foundation_NSMutableSet")]
+impl<T: Message + PartialEq> NSMutableSet<T> {
+    /// Adds a value to the set. Returns whether the value was
+    /// newly inserted.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use icrate::Foundation::{NSMutableSet, NSString};
+    ///
+    /// let mut set = NSMutableSet::new();
+    ///
+    /// assert_eq!(set.insert(NSString::from_str("one")), true);
+    /// assert_eq!(set.insert(NSString::from_str("one")), false);
+    /// assert_eq!(set.len(), 1);
+    /// ```
+    #[doc(alias = "addObject:")]
+    pub fn insert(&mut self, value: Id<T>) -> bool {
+        let contains_value = self.contains(&value);
+        // SAFETY: We've consumed ownership of the object.
+        unsafe { self.addObject(&value) };
+        !contains_value
+    }
+
+    /// Removes a value from the set. Returns whether the value was present
+    /// in the set.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use icrate::Foundation::{NSMutableSet, NSString};
+    /// use icrate::ns_string;
+    ///
+    /// let mut set = NSMutableSet::new();
+    ///
+    /// set.insert(NSString::from_str("one"));
+    /// assert_eq!(set.remove(ns_string!("one")), true);
+    /// assert_eq!(set.remove(ns_string!("one")), false);
+    /// ```
+    #[doc(alias = "removeObject:")]
+    pub fn remove(&mut self, value: &T) -> bool {
+        let contains_value = self.contains(value);
+        unsafe { self.removeObject(value) };
+        contains_value
+    }
+}
+
 unsafe impl<T: Message> Foundation::NSFastEnumeration2 for NSSet<T> {
+    type Item = T;
+}
+
+#[cfg(feature = "Foundation_NSMutableSet")]
+unsafe impl<T: Message> Foundation::NSFastEnumeration2 for NSMutableSet<T> {
     type Item = T;
 }
 
@@ -315,11 +442,41 @@ impl<'a, T: Message> IntoIterator for &'a NSSet<T> {
     }
 }
 
+#[cfg(feature = "Foundation_NSMutableSet")]
+impl<'a, T: Message> IntoIterator for &'a NSMutableSet<T> {
+    type Item = &'a T;
+    type IntoIter = Foundation::NSFastEnumerator2<'a, NSMutableSet<T>>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        use Foundation::NSFastEnumeration2;
+        self.iter_fast()
+    }
+}
+
 #[cfg(feature = "Foundation_NSEnumerator")]
 impl<T: fmt::Debug + Message> fmt::Debug for NSSet<T> {
     #[inline]
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         use Foundation::NSFastEnumeration2;
         f.debug_set().entries(self.iter_fast()).finish()
+    }
+}
+
+#[cfg(feature = "Foundation_NSMutableSet")]
+impl<T: Message + PartialEq> Extend<Id<T>> for NSMutableSet<T> {
+    fn extend<I: IntoIterator<Item = Id<T>>>(&mut self, iter: I) {
+        iter.into_iter().for_each(move |item| {
+            self.insert(item);
+        })
+    }
+}
+
+#[cfg(feature = "Foundation_NSMutableSet")]
+impl<'a, T: IsRetainable + PartialEq> Extend<&'a T> for NSMutableSet<T> {
+    fn extend<I: IntoIterator<Item = &'a T>>(&mut self, iter: I) {
+        // SAFETY: Because of the `T: IsRetainable` bound, it is safe for the
+        // set to retain the object here.
+        iter.into_iter()
+            .for_each(move |item| unsafe { self.addObject(item) })
     }
 }
